+++ src/main/java/com/team3webnovel/vo/GijeBoardVo.java
@@ -1,6 +1,8 @@
 package com.team3webnovel.vo;
 
+import java.text.SimpleDateFormat;
 import java.util.Date;
+import java.util.concurrent.TimeUnit;
 
 public class GijeBoardVo {
 	int boardId;
@@ -11,6 +13,8 @@ public class GijeBoardVo {
 	Date createdAt;
 	Date updatedAt;
 	int viewCount;
+	
+	private static final SimpleDateFormat dateFormat = new SimpleDateFormat("yyyy.MM.dd HH:mm");
 	
 	public GijeBoardVo(int boardId, int userId, String userName, String title, String content, Date createdAt, Date updatedAt,
 			int viewCount) {
@@ -79,4 +83,26 @@ public class GijeBoardVo {
 	public void setUserName(String userName) {
 		this.userName = userName;
 	}
+	
+	public String getFormattedCreatedAt() {
+        if (createdAt == null) {
+            return "";
+        }
+
+        Date now = new Date();
+        long diffInMillis = now.getTime() - createdAt.getTime();
+        long diffInSeconds = TimeUnit.MILLISECONDS.toSeconds(diffInMillis);
+        long diffInMinutes = TimeUnit.MILLISECONDS.toMinutes(diffInMillis);
+        long diffInHours = TimeUnit.MILLISECONDS.toHours(diffInMillis);
+
+        if (diffInSeconds < 60) {
+            return "1분 미만";
+        } else if (diffInMinutes < 60) {
+            return diffInMinutes + "분 전";
+        } else if (diffInHours < 24) {
+            return diffInHours + "시간 전";
+        } else {
+            return dateFormat.format(createdAt);
+        }
+    }
 }
